Removed Settings directory
