--- src/LSST/VMS/Settings/SettingReader.cpp
@@ -1,29 +1,0 @@
-/*
- * SettingReader.cpp
- *
- *  Created on: Oct 3, 2017
- *      Author: ccontaxis
- */
-
-#include <spdlog/spdlog.h>
-#include <SettingReader.h>
-
-namespace LSST {
-namespace VMS {
-
-SettingReader::SettingReader(std::string _basePath) {
-    SPDLOG_INFO("SettingReader::SettingReader(\"{}\")", _basePath);
-    basePath = _basePath;
-}
-
-const VMSApplicationSettings SettingReader::loadVMSApplicationSettings(std::string subsystem) {
-    std::string applicationSettingsFile = getBasePath("/" + subsystem + ".yaml");
-    SPDLOG_DEBUG("SettingReader: loadVMSApplicationSettings({})", applicationSettingsFile);
-    vmsApplicationSettings.load(applicationSettingsFile);
-    return vmsApplicationSettings;
-}
-
-std::string SettingReader::getBasePath(std::string file) { return this->basePath + file; }
-
-} /* namespace VMS */
-} /* namespace LSST */
